+++ frontend/src/components/Course/CourseInfo.tsx
