+++ frontend/src/components/Course/index.ts
@@ -1,0 +1,1 @@
+export { default as CourseItem } from './Course';
