--- frontend/src/components/CourseInfo/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CourseInfo';
