+++ frontend/src/components/Image/Image.tsx
@@ -5,18 +5,13 @@ import { imageDefault } from 'icons';
 import { ImageContainer } from './styled';
 
 interface Props {
-  size: string;
   imageUrl?: string;
+  width: number;
+  height: number;
 }
-const SIZES = {
-  small: 'small',
-  large: 'large',
-};
 
-const Image: React.FC<Props> = ({ imageUrl, size }) => (
-  <ImageContainer
-    style={size === SIZES.small ? { width: 150, height: 50 } : { width: 300, height: 200 }}
-  >
+const Image: React.FC<Props> = ({ imageUrl, width, height }) => (
+  <ImageContainer width={width} height={height}>
     <img src={imageUrl || imageDefault} />
   </ImageContainer>
 );
