--- frontend/src/components/Image/styled.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const ImageContainer = styled('div')({
-  overflow: 'hidden',
-  borderRadius: '10px',
-  textAlign: 'center',
-  margin: '5px',
-  float: 'left',
-});
